+++ iOS/ViewRenders/PaketSliderRender.cs
@@ -9,26 +9,33 @@ namespace PaketGlobal.iOS
 {
     class PaketSliderRender : Xamarin.Forms.Platform.iOS.SliderRenderer
     {
+        private bool isAssign = false;
+
         protected override void OnElementChanged(ElementChangedEventArgs<Slider> e)
         {
             base.OnElementChanged(e);
 
-            var slider = Control;
-            // Cast your element here
-            var element = (PaketGlobal.PaketSlider)Element;
+            if (!isAssign)
+            {
+                isAssign = true;
 
-            slider.TouchDown += (sender, args) =>
-            {
-                element.TouchDownEvent(this, EventArgs.Empty);
-            };
-            slider.TouchUpInside += (sender, args) =>
-            {
-                element.TouchUpEvent(this, EventArgs.Empty);
-            };
-            slider.TouchUpOutside += (sender, args) =>
-            {
-                element.TouchUpEvent(this, EventArgs.Empty);
-            };
+                var slider = Control;
+                // Cast your element here
+                var element = (PaketGlobal.PaketSlider)Element;
+
+                slider.TouchDown += (sender, args) =>
+                {
+                    element.TouchDownEvent(this, EventArgs.Empty);
+                };
+                slider.TouchUpInside += (sender, args) =>
+                {
+                    element.TouchUpEvent(this, EventArgs.Empty);
+                };
+                slider.TouchUpOutside += (sender, args) =>
+                {
+                    element.TouchUpEvent(this, EventArgs.Empty);
+                };
+            }
         }
     }
 }